modifice e texture

+++ src/VBO.cpp
@@ -1,7 +1,7 @@
 #include "VBO.h"
 
 // Constructor that generates a Vertex Buffer Object and links it to vertices
-VBO::VBO(GLfloat *vertices, GLsizeiptr size) noexcept {
+VBO::VBO(GLdouble *vertices, GLsizeiptr size) noexcept {
     glGenBuffers(1, &ID);
     glBindBuffer(GL_ARRAY_BUFFER, ID);
     glBufferData(GL_ARRAY_BUFFER, size, vertices, GL_STATIC_DRAW);
