+++ src/VBO.h
@@ -6,7 +6,7 @@ public:
     // Reference ID of the Vertex Buffer Object
     GLuint ID{};
     // Constructor that generates a Vertex Buffer Object and links it to vertices
-    VBO(GLfloat *vertices, GLsizeiptr size) noexcept;
+    VBO(GLdouble *vertices, GLsizeiptr size) noexcept;
 
     // Binds the VBO
     void Bind() noexcept;
